requirements.txt
requests==2.31.0
sseclient-py==1.8.0
cos-python-sdk-v5==1.9.30
tencentcloud-sdk-python==3.0.1046
